:sparkles: Added Store
Overview
===
Added Store.
Store saves state and notify new state to subscribers(receivers) when store's state is updated.
And, Store sends the new status not only when store's state renew, but also when store is subscribed.

Note
===
At first, the store contained StatePublisher and emitted state by using that StatePublisher.
But, to use StatePublisher could not ensure state update order on multi thread, because StatePublisher is not thread safe.

So, I removed StatePublisher and refactor StateRelay and to ensure Thread safe
and used StateRelay in the Store instead of StatePublisher.

The changes by refactoring made here areas follows.
1. make StateRelay to be able to publish state
2. remove StatePublisher